==> notebooks/01_load_and_validate_raw_data.ipynb <==
### Loading, exploring and validating the data

# Function to fetch data from the NYC taxi data website

from pathlib import Path
%pip install requests
import requests

def download_one_file_of_raw_data(year: int, month: int) ->Path:
    """Download one file of raw data from the NYC taxi data website."""

    url = f"https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02}.parquet"
    response = requests.get(url)

    if response.status_code == 200:
        path = f"../data/raw/rides_{year}_{month:02}.parquet"
        open(path, 'wb').write(response.content)
        return path
    else:   
        raise RuntimeError(f"Failed to download {url}")
    

# Call the function to download the data for January 2022
download_one_file_of_raw_data(2022,1)

%pip install pandas # Install pandas

import pandas as pd # Import pandas

rides = pd.read_parquet('../data/raw/rides_2022_01.parquet') # Read the data into a pandas dataframe

rides.head(10) # Show the first 10 rows of the dataframe


# Shape of the dataframe
rides.shape

# Info about the dataframe
rides.info()


**Observations and insights**

The data has 2,463,931 rows and 19 columns (January data), we may use more data in the future.

Majority of the columns are retundant for our purposes. Main columns of interest are 'tpep_pickup_datetime','tpep_dropoff_datetime','PULocationID'

We will need to generate new features through data engineering for further analysis and to improve our model in the future.

# checking for missing values
rides.isnull().sum() / len(rides) * 100

**Observations and insights**

There are a few columns with a low number of missing value (less than 3%).

As number of missing values are low, these won't affect our analysis in a meaningfull way.

Also the columns we will use for our projects do not have missing values, there is no for missing value imputation in this case.

# selecting the columns of interest for initial analysis
rides = rides[['tpep_pickup_datetime', 'PULocationID']] 

# renaming the columns for easier use
rides.rename(columns={
    'tpep_pickup_datetime': 'pickup_datetime',
    'PULocationID': 'pickup_location_id',
}, inplace=True)

rides.head() # show first 5 rows

**Conducting some sanity checks**

Lets conduct some can sanity checks to see if all loaded data is from January


rides['pickup_datetime'].describe() # show summary statistics for the pickup_datetime column

**Observations and insights**

We observe that there is a row from 2008 and a row from May 2022, let us fix this! we will take out all rows not from January 2022

rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime < '2022-02-01']
rides['pickup_datetime'].describe()

Let us save our validated data for future use.

rides.to_parquet('../data/transformed/validated_rides_2022_01.parquet') # save to parquet file
