==> notebooks/03_transform_ts_data_into_features_and_targets.ipynb <==
### Transforming the time series data into tabular data for model training

To do this we create a slidding window with slices(cutoffs). That is, for every target value, we create a vector of features. E.g for a cutoff (0,12,13) the number of features(n_feature = 12) with range(0,12), the target is range (12,13).

Below I implement this logic intially for a single location 43 to see if it works, once validated I implement this for the entire data.

import pandas as pd # import pandas library for data manipulation

ts_data = pd.read_parquet('../data/transformed/ts_data_2022_01.parquet') # load the transformed data
ts_data

# Implement this for central park (location 43)
# Let us view data for location 43
ts_data_one_location = ts_data.loc[ts_data.pickup_location_id == 43, :].reset_index(drop=True)
ts_data_one_location.head(15)

# function to create cutoff indices
def get_cutoff_indices(
    data: pd.DataFrame, # the data to be sliced
    n_features: int, # number of features
    step_size: int
    ) -> list:

        stop_position = len(data) - 1
        
        # Start the first sub-sequence at index position 0
        subseq_first_idx = 0
        subseq_mid_idx = n_features
        subseq_last_idx = n_features + 1
        indices = []
        
        while subseq_last_idx <= stop_position:
            indices.append((subseq_first_idx, subseq_mid_idx, subseq_last_idx))
            
            subseq_first_idx += step_size
            subseq_mid_idx += step_size
            subseq_last_idx += step_size

        return indices

Let us print a sample output to see if the function works.

n_features = 24 #use the whole previous day's data to generate features
step_size = 1 # slide the window by 1 hour

indices = get_cutoff_indices(
    ts_data_one_location,
    n_features,
    step_size
)
indices[:5] # view the first 5 cutoff indices

# implementing the slicing with the indices
import numpy as np # import numpy library for data manipulation

n_examples = len(indices)
x = np.ndarray(shape=(n_examples, n_features), dtype=np.float32) # features
y = np.ndarray(shape=(n_examples), dtype=np.float32) # targets
pickup_hours = []

# loop through the indices and slice the data generating the features and targets
for i, idx in enumerate(indices):
    x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
    y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
    pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

# lets see if it worked as expected
print(f'{x.shape=}')
print(f'{x=}')
print(f'{pickup_hours[:5]=}')

It works, now lets transform the array into a data frame of features for convinience.

features_one_location = pd.DataFrame(
    x,
    columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(n_features))]
)
features_one_location

Now let us do the same for the targets.

targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])
targets_one_location

Now let us write a function that does the transformation for all locations.

from tqdm import tqdm

def transform_ts_data_into_features_and_target(
    ts_data: pd.DataFrame,
    input_seq_len: int,
    step_size: int
) -> pd.DataFrame:
    """
    Slices and transposes data from time-series format into a (features, target)
    format that we can use to train Supervised ML models
    """
    assert set(ts_data.columns) == {'pickup_hour', 'rides', 'pickup_location_id'}

    location_ids = ts_data['pickup_location_id'].unique()
    features = pd.DataFrame()
    targets = pd.DataFrame()
    
    for location_id in tqdm(location_ids):
        
        # keep only ts data for this `location_id`
        ts_data_one_location = ts_data.loc[
            ts_data.pickup_location_id == location_id, 
            ['pickup_hour', 'rides']
        ]

        # pre-compute cutoff indices to split dataframe rows
        indices = get_cutoff_indices(
            ts_data_one_location,
            input_seq_len,
            step_size
        )

        # slice and transpose data into numpy arrays for features and targets
        n_examples = len(indices)
        x = np.ndarray(shape=(n_examples, input_seq_len), dtype=np.float32)
        y = np.ndarray(shape=(n_examples), dtype=np.float32)
        pickup_hours = []
        for i, idx in enumerate(indices):
            x[i, :] = ts_data_one_location.iloc[idx[0]:idx[1]]['rides'].values
            y[i] = ts_data_one_location.iloc[idx[1]:idx[2]]['rides'].values
            pickup_hours.append(ts_data_one_location.iloc[idx[1]]['pickup_hour'])

        # numpy -> pandas
        features_one_location = pd.DataFrame(
            x,
            columns=[f'rides_previous_{i+1}_hour' for i in reversed(range(input_seq_len))]
        )
        features_one_location['pickup_hour'] = pickup_hours
        features_one_location['pickup_location_id'] = location_id

        # numpy -> pandas
        targets_one_location = pd.DataFrame(y, columns=[f'target_rides_next_hour'])

        # concatenate results
        features = pd.concat([features, features_one_location])
        targets = pd.concat([targets, targets_one_location])

    features.reset_index(inplace=True, drop=True)
    targets.reset_index(inplace=True, drop=True)

    return features, targets['target_rides_next_hour']

Lets test that the function works as expected.

# Call the function to transform the data
features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*7*1, # one week of history
    step_size=24, # lets speed this up 
)

print(f'{features.shape=}')
print(f'{targets.shape=}')
