==> notebooks/04_transform_raw_data_into_features_and_targets.ipynb <==
### Automating 1,2,3: From raw data to training data.

In this notebook we combine the processes from obtaining raw data to transforming it into training data for modeling.

The functions used here are defined in the `data.py` and `paths.py` scripts.

# Telling jupyter to always reload modules before executing code
%reload_ext autoreload
%autoreload 2

Let us import 2022 data.

import sys
sys.path.append(r"C:\Users\User\capstone_project")
from src.data import load_raw_data

rides = load_raw_data(year=2022) # fetches 2022 data
rides

Let us transform the raw data into time series data.

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)
ts_data

Let us now transform the time series data into tabular data (features and targets)

from src.data import transform_ts_data_into_features_and_target

features, targets = transform_ts_data_into_features_and_target(
    ts_data,
    input_seq_len=24*28*1, # one month
    step_size=24,
)

print(f'{features.shape=}')
print(f'{targets.shape=}')

Now let us save our data to be used for modeling.

tabular_data = features
tabular_data['target_rides_next_hour'] = targets

from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')

We can now move to the next step of model building.
