==> notebooks/05_visualize_training_data.ipynb <==
### Visualizing the traing data.

Let us visualize the training data before modeling so as to: 
- further understand the data 
- check incase of any inconsistent patterns

In this notebook I use some functions from the `plot.py` script.

Loading the data

import sys
sys.path.append(r"C:\Users\User\capstone_project")


import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

features_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
features_and_target

Splitting the data into features and targets

features = features_and_target.drop(columns=['target_rides_next_hour'])
targets = features_and_target['target_rides_next_hour']

from src.plot import plot_one_sample

plot_one_sample(
    example_id=0,
    features=features,
    targets=targets,
)

**Observations and insights**

You can observe some weekly seasonality in the data, we shall use this logic in building some baseline models.

Next, let us move to the actual model building.
