==> notebooks/07_xgboost_model.ipynb <==
### Model building continued

Now, let us use machine learning to build our fourth model.

There are several option we can use for our tabular data, I will start with XGBoost a boosting tree algorithm and iterate further in next steps.

XGBoost is good for continuous variables.

### Model 4: XGBoost

Loading the training data.

import sys 
sys.path.append(r"C:\Users\User\capstone_project")

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

Splitting the data into train and test sets. Here we use the train_test_split function.

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

Importing xgboost library

import xgboost as xgb

# I use only past rides data to generate predictions. We can iterate on this later.
past_rides_columns = [c for c in X_train.columns if c.startswith('rides_')]
X_train_only_numeric = X_train[past_rides_columns]

# instantiating and training the model
model = xgb.XGBRegressor() # instantiate the model
model.fit(X_train_only_numeric, y_train) # fitting the model

# generating predictions
X_test_only_numeric = X_test[past_rides_columns]
predictions = model.predict(X_test_only_numeric)
predictions

### Evaluating model 4

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

**Observations and insights**

Model 4 performs better compared to the previous models based on MAE = 2.70

The model however takes along time to run, next let us iterate further with more models.
