==> notebooks/09_lightgbm_model_with_feature_engineering.ipynb <==
### Model building continued.

### Model 6: LightGBM with Feature engineering.

Loading the data

import sys 
sys.path.append(r"C:\Users\User\capstone_project")

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

Splitting the data into train and test set

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

Let us add a feature average number of rides for last four weeks.

def average_rides_last_4_weeks(X: pd.DataFrame) -> pd.DataFrame:
    """
    Adds one column with the average rides from
    - 7 days ago
    - 14 days ago
    - 21 days ago
    - 28 days ago
    """
    X['average_rides_last_4_weeks'] = 0.25*(
        X[f'rides_previous_{7*24}_hour'] + \
        X[f'rides_previous_{2*7*24}_hour'] + \
        X[f'rides_previous_{3*7*24}_hour'] + \
        X[f'rides_previous_{4*7*24}_hour']
    )
    return X

Let us wrap this for sklearn data transformation step.

from sklearn.preprocessing import FunctionTransformer

add_feature_average_rides_last_4_weeks = FunctionTransformer(
    average_rides_last_4_weeks, validate=False)

add_feature_average_rides_last_4_weeks.fit_transform(X_train) #apply fit_transform 

Next, let us extract hour of day and day of week from 'pickup_hour'

from sklearn.base import BaseEstimator, TransformerMixin

class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None): #fit method
        return self
    
    def transform(self, X, y=None): #transform method
        
        X_ = X.copy()
        
        # Generate numeric columns from datetime
        X_["hour"] = X_['pickup_hour'].dt.hour
        X_["day_of_week"] = X_['pickup_hour'].dt.dayofweek
        
        return X_.drop(columns=['pickup_hour']) #drop pickup_hour column to keep only numeric data

Let us check if transformation worked.

add_temporal_features = TemporalFeaturesEngineer()
add_temporal_features.fit_transform(X_train)

Let us put it al together.

import lightgbm as lgb

Putting together all steps using sklearn.pipeline.

from sklearn.pipeline import make_pipeline

pipeline = make_pipeline(
    add_feature_average_rides_last_4_weeks,
    add_temporal_features,
    lgb.LGBMRegressor()
)
pipeline.fit(X_train, y_train)

### Evaluating Model 6

The training is complete, not let us test with test data and evaluate model 6.

predictions = pipeline.predict(X_test) #predict on test set

from sklearn.metrics import mean_absolute_error
test_mae = mean_absolute_error(y_test, predictions) # calculate MAE
print(f'{test_mae=:.4f}')

**Observations and insights**

Based on the MAE, model 6 performance did not change much. 

Next,we can try hyper-parameter tuning to improve the model.
