==> notebooks/10_lightgbm_model_with_hyperparameter_tuning.ipynb <==
### Model building continued.

### Model 7: LightGBM with hyperparameter tuning

In this notebook we further improve our model with hyperparameter tuning.

import warnings
warnings.filterwarnings("ignore") # ignore warnings

Loading the data

import sys 
sys.path.append(r"C:\Users\User\capstone_project")

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

Splitting data into train and test sets.

from datetime import datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date=datetime(2022, 6, 1, 0, 0, 0),
    target_column_name='target_rides_next_hour'
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')

Defining a function which maps a set of hyperparameter to a final validation error.

%pip install optuna

import numpy as np
from sklearn.model_selection import KFold, TimeSeriesSplit
from sklearn.pipeline import make_pipeline
from sklearn.metrics import mean_absolute_error
import optuna

from src.model import get_pipeline

def objective(trial: optuna.trial.Trial) -> float:
    """
    Given a set of hyper-parameters, it trains a model and computes an average
    validation error based on a TimeSeriesSplit
    """
    # pick hyper-parameters
    hyperparams = {
        "metric": 'mae',
        "verbose": -1,
        "num_leaves": trial.suggest_int("num_leaves", 2, 256),
        "feature_fraction": trial.suggest_float("feature_fraction", 0.2, 1.0),
        "bagging_fraction": trial.suggest_float("bagging_fraction", 0.2, 1.0),
        "min_child_samples": trial.suggest_int("min_child_samples", 3, 100),   
    }
       
    tss = TimeSeriesSplit(n_splits=2)
    scores = []
    for train_index, val_index in tss.split(X_train):

        # split data for training and validation
        X_train_, X_val_ = X_train.iloc[train_index, :], X_train.iloc[val_index,:]
        y_train_, y_val_ = y_train.iloc[train_index], y_train.iloc[val_index]
        
        # train the model
        pipeline = get_pipeline(**hyperparams)
        pipeline.fit(X_train_, y_train_)
        
        # evaluate the model
        y_pred = pipeline.predict(X_val_)
        mae = mean_absolute_error(y_val_, y_pred)

        scores.append(mae)
   
    # Return the mean score
    return np.array(scores).mean()

Create a study as per optuna.

study = optuna.create_study(direction="minimize")
study.optimize(objective, n_trials=5) #run 5 iterations

Let us extract the best hyper-parameters optuna found.

best_params = study.best_trial.params
print(f'{best_params=}')

Now let us re train the model on the entire training data.

pipeline = get_pipeline(**best_params)
pipeline.fit(X_train, y_train)

### Evaluating Model 7

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)
print(f'{test_mae=:.4f}')

**Observations and insights**

Tuning hyper-parameter further improved the model based on MAE = 2.54.

This is a pretty good error, let us print to see how the predictions look like.

Printing predictions.

from src.plot import plot_one_sample

plot_one_sample(
    example_id=2979,
    features=X_test,
    targets=y_test,
    predictions=pd.Series(predictions)
)

plot_one_sample(
    example_id=3979,
    features=X_test,
    targets=y_test,
    predictions=pd.Series(predictions)
)

**Observations and insights**

As you can see, the model is relatively good at making predictions.

Model 7 is our best performing model.

Next, we can now build a batch scoring service to put the model to work.
