==> notebooks/11_backfill_feature_store.ipynb <==
### Model deployment as a batch - scoring service.

Here we use hopsworks for our feature store. In this notebook we backfill our feature store with data startin from 2022.

import sys 
sys.path.append(r"C:\Users\User\capstone_project")

HOPSWORKS_PROJECT_NAME = 'predicting_taxi_demand'

%pip install python-dotenv

import os # to access environment variables
from dotenv import load_dotenv
from src.paths import PARENT_DIR

# load key-value pairs from .env file located in the parent directory
load_dotenv(PARENT_DIR / '.env')

HOPSWORKS_API_KEY = os.environ['HOPSWORKS_API_KEY']

Download historical data from the website starting 2022 todate.

from datetime import datetime
import pandas as pd
from src.data import load_raw_data

from_year = 2022
to_year = datetime.now().year
print(f'Downloading raw data from {from_year} to {to_year}')

rides = pd.DataFrame()
for year in range(from_year, to_year+1):
    
    # download data for the whole year
    rides_one_year = load_raw_data(year)
    
    # append rows
    rides = pd.concat([rides, rides_one_year])

Check how many rows of data.

print(f'{len(rides)=:,}')

We have over 71 million rows.

Now, let us transform the data into time series data.

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)


# string to datetime
ts_data['pickup_hour'] = pd.to_datetime(ts_data['pickup_hour'], utc=True)

# add column with Unix epoch milliseconds
ts_data['pickup_ts'] = ts_data['pickup_hour'].astype('int64') // 10**6



%pip install hopsworks==3.4.*

import hopsworks

project = hopsworks.login(
    project=HOPSWORKS_PROJECT_NAME,
    api_key_value=HOPSWORKS_API_KEY
)

Now we have connected to the project.

Next let us connect to the feature store.

%pip install hsfs[python] 



import hsfs
conn = hsfs.connection()
feature_store = conn.get_feature_store()

# feature_store = project.get_feature_store()

Let us set the feature group

FEATURE_GROUP_NAME = 'time_series_hourly_feature_group'
FEATURE_GROUP_VERSION = 1

Let us create the feature group.

feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key = ['pickup_location_id', 'pickup_hour'],
    event_time='pickup_hour'
)

Inserting data into feature group.

"""
feature_group.insert(ts_data, write_options={"wait_for_job": False})
"""
